+++ Goonies/Goonies/Stalactite.h
@@ -6,7 +6,7 @@
 class Stalactite : public Entity
 {
 public:
-	void init(const glm::ivec2 &tileMapPos, ShaderProgram &shaderProgram);
+	void init(const glm::vec2 &tileMapPos, ShaderProgram &shaderProgram);
 	void update(int deltaTime);
 
 private:
